[fix] 메뉴 출력부 다시 inputView(상호작용담당)로 이동, 할일 추가로직 리팩토링
