+++ src/main/java/org/homework/func/InputView.java
@@ -1,5 +1,7 @@
 package org.homework.func;
 
+import org.homework.Options;
+
 import java.util.Scanner;
 
 public class InputView {
@@ -9,7 +11,21 @@ public class InputView {
         this.scanner = scanner;
     }
 
+    public void printOptions() {
+        Options[] options = Options.values();
+        System.out.print("옵션을 선택하세요: ");
+        for (int i = 0; i < options.length; i++) {
+            Options option = options[i];
+            String format = i < options.length - 1 ? "%s.%s, " : "%s.%s ~> ";
 
+            System.out.printf(format, option.getNum(), option.getName());
+        }
+    }
+
+    public String getNewTodoContent() {
+        System.out.print("새로운 할 일을 입력해주세요 ~> ");
+        return scanner.nextLine();
+    }
 
     public int getNum() {
         return Integer.parseInt(scanner.nextLine());
